@@ -1,26 +1,20 @@
-// Top-level build file where you can add configuration options common to all sub-projects/modules.
-
 buildscript {
-    ext.kotlin_version = '1.2.71'
+    ext.kotlin_version = '1.3.10'
     repositories {
-        maven { url 'https://maven.google.com' }
+        google()
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.2.0'
+        classpath 'com.android.tools.build:gradle:3.2.1'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-
-        // NOTE: Do not place your application dependencies here; they belong
-        // in the individual module build.gradle files
         classpath 'com.google.gms:google-services:3.0.0'
     }
 }
 
 allprojects {
     repositories {
+        google()
         jcenter()
-        maven { url 'https://maven.google.com' }
-        mavenCentral()
     }
 }
 
@@ -28,15 +22,3 @@ task clean(type: Delete) {
     delete rootProject.buildDir
 }
 
-ext {
-    buildToolsVersion = "26.0.0"
-    supportLibVersion = "26.0.0-beta2"
-    archLifecycleVersion = "1.0.0-alpha3"
-    archRoomVersion = "1.0.0-alpha3"
-    daggerVersion = "2.11"
-    firebaseVersion = "10.2.6"
-    retrofitVersion = "2.3.0"
-    okhttpVersion = "3.6.0"
-    rxAndroidVersion = "2.0.1"
-    rxJavaVersion = "2.1.0"
-}
