@@ -1,6 +1,6 @@
 package me.jaxvy.kotlinboilerplate.di
 
-import android.arch.persistence.room.Room
+import androidx.room.Room
 import android.content.Context
 import com.google.gson.Gson
 import dagger.Module
