@@ -1,10 +1,10 @@
 package me.jaxvy.kotlinboilerplate.persistence.dao
 
-import android.arch.lifecycle.LiveData
-import android.arch.persistence.room.Dao
-import android.arch.persistence.room.Insert
-import android.arch.persistence.room.OnConflictStrategy
-import android.arch.persistence.room.Query
+import androidx.lifecycle.LiveData
+import androidx.room.Dao
+import androidx.room.Insert
+import androidx.room.OnConflictStrategy
+import androidx.room.Query
 import me.jaxvy.kotlinboilerplate.persistence.entity.Item
 
 @Dao
