@@ -1,7 +1,7 @@
 package me.jaxvy.kotlinboilerplate.persistence.entity
 
-import android.arch.persistence.room.Entity
-import android.arch.persistence.room.PrimaryKey
+import androidx.room.Entity
+import androidx.room.PrimaryKey
 
 @Entity
 class Item {
